# Combine tallying of buckets for both passes of float32 sorting