The `buckets16` array is re-used between the two passes. However, the tallying step counting to which bucket each splat belongs is independent. By using separate arrays the tallying can happen in one loop for both passes. This does require more memory, but saves time.

Here's a table with the results I've got using the `benchmarkSort` method. There is only one set of columns for the non 32float sorting as it's unaffected by this change and only provided for context:
| Browser | JS | WASM | JS32 before | WASM32 before | JS32 after | WASM32 after |
|--|--|--|--|--|--|--|
|Firefox|1.59 ms | 1.75ms | 6.79 ms | 6.89 ms | 5.55 ms | 5.35 ms |
|Chrome| 2.32 ms | 1.53 ms | 10.86 ms | 5.53 ms | 10.12 ms | 4.30 ms |